Check if any keys are in the agent before running ssh-add

+++ .bash_profile
@@ -5,9 +5,10 @@ export LANG=en_GB.utf8
 export PAGER=less
 export PATH=/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin
 
-if ! pgrep -u "$EUID" -x ssh-agent >/dev/null 2>&1; then
-    eval "$(ssh-agent | tee ~/.ssh-agent)" >/dev/null
+[[ -r ~/.ssh-agent ]] && . ~/.ssh-agent >/dev/null
+if ! ssh-add -l >/dev/null 2>&1; then
+    if ! pgrep -u "$EUID" -x ssh-agent >/dev/null 2>&1; then
+        eval "$(ssh-agent | tee ~/.ssh-agent)" >/dev/null
+    fi
     ssh-add
-else
-    [[ -r ~/.ssh-agent ]] && . ~/.ssh-agent >/dev/null
 fi
